Move ActiveWorkspace to separate set of files
