+++ caffe2/core/workspace.cc
@@ -312,8 +312,4 @@ ThreadPool* Workspace::GetThreadPool() {
   return thread_pool_.get();
 }
 
-std::mutex ActiveWorkspace::wsmutex_;
-
-std::unordered_set<Workspace*> ActiveWorkspace::workspaces_;
-
 } // namespace caffe2
