+++ caffe2/core/workspace.h
@@ -300,46 +300,6 @@ class Workspace {
   DISABLE_COPY_AND_ASSIGN(Workspace);
 };
 
-/**
- * Captures and provides global access to a set of active workspaces in a
- * thread-safe fashion.
- */
-class ActiveWorkspace {
- public:
-  /**
-   * Borrows a workspace and adds it to the active set.
-   */
-  explicit ActiveWorkspace(Workspace* workspace) : workspace_(workspace) {
-    CHECK_NOTNULL(workspace);
-    std::lock_guard<std::mutex> guard(wsmutex_);
-    auto inserted = workspaces_.insert(workspace_).second;
-    CHECK(inserted) << "Workspace is already borrowed as active!";
-  }
-
-  ~ActiveWorkspace() {
-    std::lock_guard<std::mutex> guard(wsmutex_);
-    workspaces_.erase(workspace_);
-  }
-
-  /**
-   * Applies a function f to each active workspace.
-   */
-  template <typename F>
-  static void ForEach(F f) {
-    std::lock_guard<std::mutex> guard(wsmutex_);
-    for (Workspace* ws : workspaces_) {
-      f(ws);
-    }
-  }
-
- private:
-  Workspace* workspace_;
-
-  static std::mutex wsmutex_;
-
-  static std::unordered_set<Workspace*> workspaces_;
-};
-
 }  // namespace caffe2
 
 #endif  // CAFFE2_CORE_WORKSPACE_H_
